Goodbye to Subversion, we're now at GitHub. Fiat lux!
After almost three years of hectic software development, it is finally
time to say goodbye to Subversion and embrace the distributed revision
control system paradigm. This starts as a clean repository, continuing
from the point where we were with SVN. Trust me, there is not much you
would find it useful to take a look at in the old repository.

[We could have cleanly moved our Subversion repository and all its
history to GitHub, but, truth be told, it suffered from many, many
common greenhorn mistakes that there is no need to expose to the world
and which would no nothing but drag our names through the mud]