FITSet.find_fits replaced with fitsimage.find_files.
The fitsimage.FITSet.find_fits method had the problem that, as it
detected FITS images as directories were walked down, it was not
possible to have an estimate of how long the process of finding the
images was going to take. Also, instead of sequentially checking which
files are FITS images, those modules using the multiprocessing module
to do their job could also perform this verification in parallel.

Thus, the module-level method fitsimage.find_files has been added. It
also walks-down the input paths, as FITSet.find_fits did, but only to
obtain a list of all the regular files there located. This is a fast
operation, as files need only be listed. This list is then distributed
among the workers (offsets.py) or its length used to display a
progress bar as FITS files are detected (import.py).