Add --exact option, to not modify the FITS images.
If this option is used, an exact copy of the FITS images is imported:
they are only detected, copied and renamed sequentially. Thus, FITS
headers are not modified — in particular, this means that the --uik
keyword (which stores the path to the original file, so that it may
be accessed later by other modules) is not added.

The SHA-1 cryptographic hash function is used to verify that the copy
of the FITS images is identical, raising an error otherwise.