Add run_tests.py — execute all unit tests at once.
This is a convenience script for loading all the unit tests in the
test/ directory and running them. Test modules are automatically
detected: those files which start with 'test_' and have the '.py'
extension (such as test/test_passband.py) are treated as such.

For Python 2.6 compatibility, we cannot use the 'failfast' option,
which causes the test run to stop on the first error or failure.