Make LEMONdB.add_pfilter private.
Prefix an underscore to the method name (_add_pfilter) and make
LEMONdB.add_image and LEMONdB.set_rimage invoke it directly. In this
manner, photometric filters are automatically added when images are
inserted into the database, and we release ourselves from the burden
of having to do it manually at the photometry module.