Add support for SAVEPOINTs to LEMONdB.
SAVEPOINTs are a method of creating transactions, similar to BEGIN and
COMMIT, except that SAVEPOINT and RELEASE commands are named and may
be nested.

You may think of savepoints as 'marks' in the transaction timeline.
In this view, the SAVEPOINT command creates a new mark, the ROLLBACK
TO command rewinds the timeline back to a point just after the named
mark, and the RELEASE command erases marks from the timeline without
actually making any changes to the database.

http://www.sqlite.org/lang_savepoint.html