Make LEMONdB.add_image raise our own exception (DuplicateImageError)
Instead of sqlite3.IntegrityError, raise DuplicateImageError when we
attempt to add to the database an image whose observation date is the
same than that of another already-stored image.

Also, use a savepoint to roll back the insertion of the filter and
photometric parameters (as, internally, LEMON._add_pfilter and
_add_pparams are called before the image is added) of the image when
DuplicateImageError is raised. In this manner, the database is not
unnecessarily modified (adding rows to the PHOTOMETRIC_FILTERS and
PHOTOMETRIC_PARAMETERS tables) if the insertion cannot be completed.