Make LEMON.get_star_info raise KeyError instead of returning None.
If the specified ID does not match that of any other star stored in
the database, raise KeyError instead of returning None. This is much
less error-prone, particularly when our code should never attempt to
fetch a star whose ID is not in the database -- that is always an
error, so we better directly raise an exception.