Rename LEMONdB.get_star_info and LEMONdB.get_star_info methods.
LEMON.dB.get_star, being the logical opposite of add_photometry (the
latter inserts a photometric record, while the former returns a DBStar
with all the photometric information) becomes LEMONdB.get_photometry.

In turn, LEMONdB.get_star_info becomes get_star, as it is the opposite
of LEMONdB.add_star: one inserts the coordinates and magnitude of a
star into the database, and the other extracts it.