Add LEMONdB._star_pfilters method.
Using a single query to extract the photometric filters for which any
star in the database has information, as LEMONdB.pfilters does, has
the disadvantage that it does not allow us to check in which filters a
particular star was observed. Although not critical, this information
is useful in order to be able to raise the appropriate error messages
in other methods of the LEMONdB class.