Make LEMONdB._add_cmp_star raise our own exceptions.
Three are the different exceptions that the method may raise: KeyError
if any of the two stars is not in the database, ValueError if the IDs
of the two stars are equal (as a star cannot use itself as comparison)
and UnknownFilterError if there are no photometric records in the
database for any of the stars involved.

To rephrase the last sentence: in order to add a comparison star, both
stars must have been observed (and stored with LEMONdB.add_photometry)
in that photometric filter. This means that, in order to save a light
curve in the Johnson V filter, for example, the photometric records of
both stars in Johnson V will have to have been added first.