Remove UnknownFilterError from LEMONdB._add_cmp_star method.
After having given some thought to it, we're going back on our word.
Forcing the user to add the photometric records (LEMON.add_photometry)
of the stars before a light curve can be stored may render the class
useless in case we need to work with a light curve computed elsewhere,
whose original photometric data could be unavailable.

As the photometric filter may have to be added to the database, use a
savepoint to roll back its insertion (LEMON._add_pfilter) in case any
of the remaining exceptions is raised because of an error. This leaves
the database in its original state (without unnecessarily adding a row
to the PHOTOMETRIC_FILTERS tables) if the insertion fails.