Change signature of LEMONdB.add_period method.
Modify the order of the last two arguments of the method, so that the
period of the star is passed before the step that was used to compute
it using the string-length technique. The step is used to estimate how
reliable the period is, whereas the real important data is the period,
so it makes sense to put the latter first.