Make LEMONdB.add_period raise our own exceptions.
Instead of sqlite3.IntegrityError, raise UnknownStarError in case the
ID of the star does not match that of any of the stars stored in the
database, and DuplicatePeriodError if we attempt to add the period of
the same star twice.

As the photometric filter must be added if it is new to the database,
the modification is made atomically, so the LEMONdB is guaranteed to
be left untouched in case an error is encountered.