Modify FITSeingImage class to use the real saturation levels.
Thus far, the saturation level used when detecting sources with
SExtractor has been that defined in the sextractor.sex configuration
file (SATUR_LEVEL parameter). While this simple approach sufficed
during the early phases of development, we need our code to reduce
data taken with different instruments and, more important, also be
able to work with coadded images.

In the latter scenario, the number of ADUs at which arises saturation
depends on the number of images that were coadded, but SExtractor
cannot take this into account. As a result, all the stars can be
marked as saturated as even the sky background may be tens of times
higher than the value of SATUR_VALUE in the configuration file.

Although we do not use SExtractor for photometry, but only for sources
detection, knowing which stars are in actuality above the saturation
level would allow us to do things such as, when computing the median
of the FWHM of the stars in an image, exclude those saturated. It may
also improve the calculation of the offset between images, as the
information on the center of saturated objects is not very reliable.

Update FITSeeingImage's __init__ so that it receives the CCD
saturation level and the FITS keyword which stores the number of
images that were coadded. The effective saturation level is that which
results from multiplying the saturation level by the number of coadded
images. For example, if the CCD saturates at 55,000 ADUs and 25 images
are coadded, the effective saturation level is 55,000 x 25 = 1,375,000
ADUs. If the FITS keyword cannot be found, a value of one is assumed.

This modification requires also to change how the MD5 hash of the
SExtractor configuration files is used in order to cache on disk the
catalog and avoid having to unnecessarily compute it again. The
checksum must also include the -SATUR_LEVEL option and its value,
that overrides the definition in the SExtractor configuration file,
since it can now be different in each instantiation of the class.

Also, take advantage of the refactoring to simplify the code of
qphot.QPhot and make it mandatory to pass to it the sequence of
coordinates on which photometry is to be done. The feature of doing
photometry on the sources automatically detected by SExtractor has
never been used and will not be either in the foreseeable future.