Fix bug in Star.angular_distance method.
The formula used to compute the angular distance between two Stars
from their coordinates is correct, but instead of using their right
ascension and declination the code takes their x- and y-coordinates
in the image. This seems to be the result of a careless refactoring in
the past, when the angular_distance method was moved from the Pixel to
the Star class.