Attempt to improve the precision of the astrometric calibration.
Add two of the optional SExtractor catalog fields that are used by
SCAMP to filter out blended and corrupted detections, as well as to
identify small glitches or extended objects, and which in theory
should help improve the precision of the calibration. Tweak also some
parameters in the SCAMP configuration file with this purpose in mind.