Modify the default value of --uik option
Use 'UNCIMG' as the default FITS keyword which stores the path to the
uncalibrated image. The problem with "LEMON UNCALIBRATED IMAGE", the
current one, is that PyFITS does not support CONTINUE for HIERARCH
FITS keywords, so it complains that its value is too long (raising
ValueError) when we exceed 80 characters.