Modify the behavior of the _exclude_index keyword argument
The current implementation of StarSet.light_curve requires that, when
the _exclude_index argument is used, the number of weights passed to
the method be equal to that of stars in the set minus one. Then,
internally, a zero is inserted in the position of the excluded star.

This can be simplified and the usage of StarSet.light_curve made
simpler if the method takes care of overriding the value of the
_exclude_index-th coefficient, assigning to it a value of zero, and
rescales the others. In this manner, it is not necessary to manually
remove the coefficient of the star to be excluded before calling the
method, as we are currently forced to do in StarSet.broeg_weights.