Discard FITS images with duplicate observation dates
Until now, photometry.py aborted the execution when it found two or
more images with the same observation date. Although a rare event,
most of the times this happens it is caused by negligence from the
user's side, such as making in the input directory a copy of one
of the FITS images that will be reduced by the pipeline.

However, astronomical instruments also have software errors, and we
have already encountered this situation while reducing images taken
with Omega2000, a wide-field camera for the 3.5m telescope at Calar
Alto: two images, although different in all other aspects, had the
same observation date. Not to pin the blame on somebody else, but
apparently something went wrong at the observatory and an incorrect
observation date was written to one of the FITS headers.

And that is precisely the issue: if we come across two images with the
same observation dates, which one has the right date? There is no way
we can find this out. Even worse, it may turn out that both images
have the wrong date; after all, something did already go wrong, for
what we know, and it could be worse than what we think.

The reason why we cannot have multiple images with the same date is
that light curves, which are our ultimate goal here, can only have one
differential magnitude for each point in time. This restriction shall
be partially lifted the day our pipeline can combine images taken with
different instruments, which is not currently the case. When that time
comes, duplicate observation dates should be fine as long as images
were observed with different filters.