First version of the graphical user interface for data analysis.
Except for the About window, the GUI is not functional; it just shows
a very few menu items and buttons whose signals are not even connected
to their callbacks. This is just a minimal, barely working example of
how the interface with which light curves will be analyzed should,
roughly, look like.

A word of caution: this is my first foray not only into PyGTK, but
also GUI programming at all. I'm using the excellent slides and code
examples that Juan Pedro Bolívar kindly shared with us last year
during the course given at the Institute of Astrophysics of Andalusia
(http://www.sinusoid.es/python-avanzado), but still it is a certainty
that, as someone with absolutely no previous experience with this,
many unforgivable bad practices will be perpetrated.