Show a dialog box with a progress bar when a LEMONdB is loaded
Loading a LEMON database into memory is not immediate, and may easily
take several seconds, so we use a progress bar to inform the user of
the status of the operation. A 'Cancel' button makes it possible to
abort the loading. When the process is complete, a second dialog box
lets the user know about how many stars were loaded.