Make the list of stars sortable by any column
Note that sexagesimal coordinates, stored as strings, are not sorted
lexicographically, but instead according to the values in the column
with the decimal coordinates. This is so because, in alphabetical
order, "+09d 23m 59.63s" precedes "-04d 29m 43.78s", for example,
while we would expect the latter (-4.495 deg) to be listed before
the former (9.4 deg).