Select the unit in which star periods are shown
Implement the functionality of the View — Coordinates submenu, which
allows to choose among three different formats: days (0.900694, for
example), hh:mm:ss (21:37:00) and seconds (77820).

Internally, this is implemented by creating three columns, with the
different formats, for each period, and displaying at a time only the
one that is selected. The columns with the periods in hh:mm:ss are
sorted according to their value in seconds, not lexicographically.

Until we figure out how to use empty cells for those photometric
filters for which the period of a star is unknown, -1 is used.