Allow only to have one LEMONdB open simultaneously
If the user attempts to open a second LEMONdB, display a dialog box
warning that data analysis must be done on one LEMONdB at a time and
therefore the current tabs will be automatically closed unless the
operation is canceled.

Imposing this restriction, at least for now, seems to be preferable to
increasing the complexity of the GUI, as otherwise the user may end up
with multiple tabs with stars belonging to different LEMONdBs.