Update the progress bar only when the percentage varies
Do not update the progress bar for each star of the LEMONdB that is
loaded into memory, but instead only when the percentage changes. If,
for example, it is 0.971 (97%), setting it to 0.972 (still 97%), and
particularly rendering it again, when the next star is loaded means
that the process is unnecessarily slowed down.