Open a page with the details of a star when double-clicking on it
These pages allow to select one of the photometric filters for which
the star has information, plotting (a) its light curve, (b) the list
of comparison stars, and their weights, that were used to compute it
and (c) the detailed list of points of the light curve. The methods
used here are only stubs, and even the layout of the page is just a
crude draft, so this should only be considered a first approximation.