Display the detailed list of light curve points
When a photometric filter is selected, show a second GtkTreeView with
the data for each point of the light curve: (a) time of observation,
(b) differential magnitude, (c) signal-to-noise ratio and (d) maximum
and minimum errors, in magnitudes, induced by noise. By default, these
rows are sorted by the date (first column).