Plot airmasses in the light curves
Add a check button which controls whether airmasses are shown in the
light curves. When enabled, airmassed are plotted on a second y-axis,
connecting with blue dashed lines the airmasses of the points of the
curve that are fewer than 'delta' seconds apart. In this manner,
airmasses from different nights are prevented from being connected by
the same line, which would not add any useful information and would
only clutter the plot unnecessarily.

This first version uses a hard-coded delta of 10800 seconds (three
hours), which works more than acceptably well with our data, until
the option of specifying its value is implemented.