Make the 'delta' of airmasses customizable
Use a spin button to allow the user to select the 'delta' value with
which airmasses are plotted. It defaults to three hours, and is
automatically disabled (made insensitive in GTK parlance) when the
option of plotting airmasses is unchecked.

In this first implementation, every time the user selects a new value
in the spin button the light curve must be plotted again. This means
that using the arrows to increment or decrement the value of 'delta'
is extremely slow, as the plot is redrawn many times. Future versions
should improve this be delaying the redrawing until the user presses
some button, or, maybe even better, until a few seconds have passed.

Note that the airmasses check button and the 'delta' spin button are
independent for each star. This should be enough for now, but at some
point these two options should be made global, and maybe even moved to
a drop down menu in the menu bar.