Show full dates in interactive navigation
When datetime instances are plotted in matplotlib and we hover the
mouse over the window, the x-coordinates are formatted (at the lower
right corner of the toolbar) the same say the tick labels are, such as
"Jan 04 2012". We can use one of the built-in date formatters of
matplotlib to achieve a higher degree of precision, so that dates are
displayed with the same format as time.ctime(): "Www Mmm dd hh:mm:ss
yyyy", where Www is the weekday, Mmm the month in letters, dd the day
of the month, hh:mm:ss the time, and yyyy the year.