Add functionality for the Close button
The LEMONJuicerGUI.handle_close method removes the current page in the
gtk.Notebook. If the page contains the details of one of the stars, it
is removed without any complaints. However, closing the page with the
list of stars of the LEMONdB (the first one) is interpreted as having
finished with it, and thus all the pages in the notebook are removed.
A dialog box asks the user to confirm this action.

If no other page is open (that is, if there is only one page in the
notebook, that with the list of stars), no confirmation is required.