Prevent dates from overlapping on the x-axis
Matplotlib allows us to natively plot datetime instances. Although for
the most part it does a good job picking tick locations and string
formats, these are sometimes squashed together. In order to avoid
this, we can set the maximum number of x-axis ticks to eight, a value
which seems to work well with our minimum width of 768 pixels.

Note that the maximum number of ticks ('maxticks' keyword argument in
matplotlib.dates.AutoDateLocator.__init__) was not available until
version 1.0.0 (July 2010). Nothing has been implemented in our code
to deal with older versions, so dates may overlap in those cases.