Show the main information of the star
Add a third gtk.TreeView to the page with the details of each star,
using it to display the same information that is shown in the list
of stars of the LEMONdB: image and sky coordinates, instrumental
magnitude in the reference image and the period in the different
photometric filters.

The headers of the TreeView are not visible; we just need a simple
table of two columns, where the first has the name of the attribute
(such as "Right ascension") and the second the corresponding value
("06h 41m 04.41s", for example). Showing the generic headers
"Attribute" and "Value" would not add any useful information.

This first version does not yet take into account the options selected
in the View menu. Therefore, coordinates are shown in both sexagesimal
and decimal formats, and each period in three different units: days,
hh:mm:ss and seconds.