Select the unit of the periods also when star details are shown
Implement the functionality of the View — Coordinates submenu in the
StarDetailsGUI class too, so that three different formats: days
(0.8889 days, for example), hh:mm:ss (21:20:00) and seconds (76800)
can be chosen.

In order to be able to hide some of the rows of a TreeView, we need to
use a TreeModelFilter. This class acts as a wrapper for our TreeModel
(a ListStore or TreeStore), allowing us to decide which rows are
displayed based on the value of a column with boolean values. Thus, we
need to add a third and last column to the store, which controls the
visibility of the period rows and whose values are updated when a
different unit is selected.