Do not use subprocess.check_call to run photometry.py
Now that the PyRAF process-caching memory is turned off (see commit
b7c650d174), photometry.main() can be directly invoked. We were using
the subprocess module as a workaround for what we erroneously thought
was a bug in PyRAF, running photometry.py separately so that it did
not 'corrupt', as it seemed, our Python namespace.