Require PyRAF version 2.0 or newer
Until now, once in a while IRAF's imexpr would incomprehensibly raise
an exception (KeyError: "Error in parameter 'a' for task imexpr\n'Key
a not found'"), although the 'a' operand referenced by the expression
had indeed been defined, and therefore the execution was aborted with
the now-familiar "PANIC in `/iraf/iraf/bin.linux/x_images.e': Write to
IPC with no reader" error message.

This started occurring when the code was parallelized using the
multiprocessing module and happened only with pools of at least two
workers, but other than that the error was not deterministic: it would
sometimes take just a few seconds to appear, while most of the times
it would not arise until well into the execution of the program.

This was verified to happen on several machines with different
versions of both IRAF (2.14 and 2.16) and PyRAF (1.8 and 1.9), reason
by which we were almost positive that there was something wrong with
PyRAF — until César Husillos, praise be to him, found out that the
error did not occur with the latest version of PyRAF, that is, 2.0.
It seems that some of the changes in this version fixed the issue,
although we have not found anything in the release notes about it.