Print updates of the progress bar to a new line when debugging
Progress bars are implemented using carriage returns: they start with
a '\r' character, thus causing all text to be echoed at the beginning
of the line every time they are printed. However, this means that,
when the logging level is set to INFO or lower, the first logging
message between two updates of the progress bar is appended to the
text of the first one — which is ugly. We can fix this by printing
the progress bar to a new line each time.