Use only one thread for each instance of SExtractor
By default, when run on symmetric multiprocessor systems, SExtractor
automatically determines the number of threads to be used in order to
speed up the execution. However, the seeing.py module runs SExtractor
in parallel, spawning multiple instances of it, so we better restrict
the number of threads to one per process.