Make the 'Plot airmasses' option global
Using an independent checkbox to determine whether the airmasses of
each star are plotted gives excessively detailed control and ends up
being even confusing — especially taking into account that the number
of hours used for the intervals is also set separately. Normally the
user is only interested in showing or hiding the airmasses from the
plots all at once and without further consideration.

Thus, we have decided to move the 'Plot airmasses' checkbox to the
View menu. The length of the intervals cannot yet be chosen: until
this is implemented we are using a fixed value of three hours, which
should work fine in almost all conceivable scenarios.