Show magnitudes and standard deviations of reference stars
In the list of comparison stars used to compute each light curve,
include also their instrumental magnitude and standard deviation.
This information, as Matilde Fernández has explained, is useful to
evaluate at a glance the precision of the photometric measurements
of both the star under study and those used as comparison.