Migrate XMLOffsetFile from xml.dom.minidom to lxml
The Python light-weight implementation of the Document Object Model
interface, simple and convenient enough in the early stages of our
development, scales awfully with the size of the XML files. Let us
note, for illustration purposes, that a 2.2-megabyte XML file with
6324 offsets took this morning almost twenty minutes to parse.

I will not argue that there had to be room for improvement in our
implementation using xml.dom.minidom, but after the refactoring, now
with lxml, the time needed has been reduced down to ~1.3 seconds. That
is more than two orders of magnitude of speed-up. It really is, as it
fancies itself to be, outstandingly fast!