Catch exception when there are no light curves for an aperture
The method mining.LEMONdBMiner.sort_by_curve_stdev raises
NoStarsSelectedError in case nothing can be returned (that is, if the
light curve of none of the stars in the specified photometric filter
has the indicated minimum number of points). This exception was not
being catched, so, although a rare event, the script would abort if
this happened when computing the median of the standard deviation of
the light curves for a specific aperture and sky annulus. The fix is
as simple as catching the exception and moving to the next iteration
of the loop.