Include debug messages throughout FITSeeingImage.__init__
There already were some DEBUG and INFO messages, but now the method
keeps track of all the events that occur. FITSImage.update_keyword,
which is called by FITSeeingImage.__init__, has also been modified to
include the appropriate log messages — it is the first of the many
methods of the FITSImage class that will make use the logging module.