Use 'sex' as an alternative name for the SExtractor executable
The name of the SExtractor executable is 'sex', but in the Debian
package it was renamed to 'sextractor', which is what we have used
(and therefore also expected) in our code. We have realized this now
because we have had to install SExtractor on a Red Hat Enterprise
Linux system, for which we have downloaded for the first time the RPM
package available at the Astromatic website. We now take into account
that the SExtractor executable could have any of these two names.