Set SATUR_KEY parameter to a non-default value
The SATUR_KEY parameter, which defaults to 'SATURATE', sets the FITS
keyword from which the saturation level, in ADUs, is read. If present
in the image header, the value of this keyword overrides, as Emmanuel
Bertin himself has amiably explained to us, the one given in the
SATUR_LEVEL parameter.

This is precisely what was causing SExtractor to work strangely with
one of our images: because of a 'SATURATE' keyword with a value of 5
ADUs, all the stars were being flagged as saturated regardless of the
value that we gave as saturation level. To avoid this happening again
in the future, we must set SATUR_KEY to a value with which we should
never conceivably come across.