Save settings to (and read them from) a configuration file
Instead of using the default options (those defined in the Glade XML
file, such as whether coordinates are expressed in sexagesimal format,
decimal or both) every time the program starts up, write them to disk
so that they persist across different executions. In this manner, the
options selected by the user (for now, only in the View submenu) are
always the same and do not have to be changed again the next time.

In order to implement this feature, we subclass SafeConfigParser and
have the destructor automatically update the configuration file when
the instance is about to be destroyed. For this to work, of course, we
need to update the SafeConfigParser as soon as any option is modified,
as well as, now that we are stating the obvious, set them at startup
to the values specified in the configuration file.

This first version uses 'juicerc' as the name of the configuration
file, expecting it to be in the current working directory. If it is
not found, the file is created with default values.