Plot light curves in each photometric filter with a different color
Matplotlib has so far automatically chosen the color with which light
curves are drawn — when undefined, the first element in the cycle of
colors is used. After a brief meeting with Matilde Fernández and very
little discussion, we have reached consensus on which colors should be
used for each of the thirteen photometric filters currently supported:

Letter |   Color
-------------------
   U   | Violet
   B   | Blue
   V   | Green
   R   | Light red
   I   | Dark red
   Z   | Cyan
   Y   | Brown
   J   | Yellow
   H   | Pink
  K(s) | Orange
   L   | Light gray
   M   | Dark gray

These are the default colors, used for both the circle markers and the
vertical errorbars in the light curves. Although they cannot (yet) be
modified through the graphical interface, they can be edited manually
in the configuration file.