Include standard deviations in the overview of a LEMONdB
In the list of stars that is presented to the user when a LEMONdB is
loaded, add a new column for each photometric filter, containing the
standard deviation of the corresponding light curve. Until we figure
out how to use empty cells for those photometric filters for which the
light curve of a star, and thus its standard deviation, is unknown, -1
is used.

Note that this slows down quite considerably the process of loading a
LEMON database into memory, as for each star and photometric filter we
need to extract the points of its light curve and compute the standard
deviation of the differential magnitudes.