Add the standard deviations to the detailed view of each star
Include in the gtk.TreeView with the details of each star the standard
deviation of its light curves. As it is the case with the overview of
the LEMONdB (which includes this information since commit 43c52a0d4e),
-1 is temporarily used when the light curve in a photometric filter
could not be generated and thus the standard deviation is unknown.