Simplify the method signature of StarDetailsGUI.__init__
When StarDetailsGUI is instantiated, it receives three parameters
(gtk.Builder, config.Configuration and database.LEMONdB) which must be
shared with it by the parent LEMONJuicerGUI class — as we need to have
access to the same user interface, configuration settings and database
handler. However, since these three parameters are already attributes
of each LEMONJuicerGUI instance, the code can be simplified by making
StarDetailsGUI.__init__ receive a reference to it.