Fix bug in LEMONJuicerGUI.handle_close
The list of open stars is not updated when a tab is closed, so, if we
click again on a star that was previously opened and closed, the GUI
attempts to switch to a nonexistent page of the notebook. The fix is
as easy as to update the list of open stars before removing a page.