Remain on the same filter when double-clicking on a reference star
If the user double-clicks on one of the reference stars, we want to
open or switch to it displaying its light curve in the same filter
from which we come. The reasoning behind this is that, most of the
time, if we are studying a light curve in one photometric filter (say,
Johnson V) and double-click on one of its reference stars, we want to
stay in the same photometric filter — instead of, as it has been done
so far, opening it by its first filter (by wavelength, which might be,
say, Johnson B) or just switching to it and thus display the last
filter that was selected by the user.

Note that this new functionality does not apply when the user clicks
on a star in the overview of the LEMONdB: in this case there is no
photometric filter we come from, so we open the star or, if already
open, switch to it showing the data for its first photometric filter
or the last one that was previously selected, respectively.