Redirect standard error of 'which' to the null device
The 'which' command that comes with GNU/Debian, the operating system
we are using for development, does not print anything in case the
specified command is nonexistent or not executable — just exiting with
return code 1. However, the version bundled with Red Hat Enterprise
Linux, we have found out, prints PATH to standard error when the
command cannot be located. We need to capture this stream too, as,
either if 'which' succeeds or fails, we want it to run silently.