Disable photometric filters for which there is no light curve
Each StarDetailsGUI displays a button for each of the photometric
filters for which there is information in the LEMONdB, but it may
happen that the light curve of a star could not be computed in one or
more of these filters. Therefore, clicking on these buttons raises an
error as we are attempting to plot a light curve which does not exist.

In order to fix this, StarDetailsGUI.__init__ must set to insensitive
(i.e., gray out) the buttons corresponding to those filters for which
the light curve of the star does not exist. This also implies that,
when we double-click on a star in the overview of a LEMONdB, it must
open at the first filter, when sorted by wavelength, among those for
which the light curve of the star is known.