Replace -1's for unknown star periods with empty cells
The visibility of the columns in a gtk.TreeViewColumn is determined by
the 'visible' keyword argument of its constructor, which specifies the
tree model column from which the attribute of the same name will be
retrieved. The cell is thus displayed only when the attribute is True.

This is exactly what we can use to get rid of those -1 values that, we
always knew it was temporary, we have been using when the period of a
star is unknown — simply leaving the cell empty is not an option, as
the values in a gtk.ListStore must always be convertible to the type
of the column. But if instead of -1 we use 0, the 'visible' attribute
will evaluate to False and thus the cell will not be shown.

Therefore, we now use the same column by which the gtk.TreeViewColumn
is sorted to also establish its visibility. This has required a minor,
additional change: as we are using the same index for both properties,
columns with the periods in days (a real number) must be sorted by the
column with the period in seconds (an integer). Otherwise, we would
get the "unable to set property `visible' of type `gboolean' from
value of type `gdouble'" error message — in other words: PyGTK seems
not to be capable of determining the truth value of a real number.