Use empty cells for unknown standard deviations too
The previous commit, 0516d80c, replaced -1's in the columns with the
periods of the stars with empty cells, but we are still using these
values for the corresponding standard deviations: if the light curve
could not be generated there are no differential magnitudes whose
standard deviation can be calculated either. Instead of a mediocre
-1, we also want to display here an empty cell.

To establish the visibility of these cells, which store floats, we
require an additional column for each standard deviation, as PyGTK
apparently cannot determine the 'visible' attribute from the truth
value of real numbers. These second columns, one per photometric
filter, store a boolean value indicating whether the corresponding
standard deviation must be shown — and are, of course, for internal
use, and thus hidden from the user.