Hide unknown standard deviations in detailed view of stars
Instead of the -1 values that we have, as a temporary measure, used
when the light curve of the star could not be generated and therefore
the corresponding standard deviation is unknown, hide those rows in
the gtk.TreeView with the details of each star. They still exist, now
using the module-level variable UNKNOWN_VALUE, but are not visible.