Add message window with the parameters of the first search algorithm
This is the first algorithm for data analysis that, after our last
meeting with Matilde Fernández, we are going to implement. This commit
does not add any real functionality other than the dialog window which
allows to configure the parameters of the search. It has also been
added a new submenu, 'Find', where the first data analysis algorithms
that we have in mind will be placed.

What this search algorithm will do is allow the astronomer to identify
those stars whose light curve amplitudes increase (on decrease, as the
user prefers) with the wavelength. For example, if the photometric
filters of the light curves were Johnson B, V and I, the stars found
by this method would be those whose A{B} <= A{V} <= A{I} (or the other
way around, if the user chooses so), where A{x} represents the
amplitude of the light curve in the photometric filter 'x'.

The amplitude of a light curve is defined as the difference between
its maximum and minimum, but we are taking into account that there may
be outliers, product of measurement errors, that could significantly
distort the result. In order to prevent this, amplitudes are defined
as the difference between the mean or median of the maximum and
minimum values in the light curves.

An option for excluding from the search stars with 'noisy' amplitudes
has also been included. To determine whether the amplitude of a star
is noisy, we calculate how many times it is greater than the mean or
median of the standard deviations of the light curves of other stars
with the most similar brightness. If this ratio is smaller than the
threshold selected by the user, the amplitude is marked as noisy and
the star is discarded.

Note that, in the dialog window, the radio buttons adjacent to the
"Amplitudes must be sorted by" text have the labels "BVI" and "IVB".
These are temporary values: in subsequent commits they should be
updated dynamically to reflect the actual photometric filters for
which there is information in the LEMONdB.