Encapsulate the message window in its own class
This may not be necessary for simple message boxes, but the window
with the parameters of the search algorithm will not only allow the
user to adjust them: more complex functionality (such as disabling
those that are not applicable or transparently interacting with the
configuration file) is expected. The way to go, therefore, is to
move the window into its own class, AmplitudesSearchMessageWindow.